+++ lib/data/models/vocabulary.dart
@@ -1,7 +1,7 @@
-import 'dart:ffi';
+
 
 class Vocabulary {
-      Long id;
+      int id;
       String word;
       String meaning;
     Vocabulary({required this.id,required this.word,required this.meaning});
